% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_hg19DBNM.R
\docType{data}
\name{hg19DBNM}
\alias{hg19DBNM}
\title{hg19-RefSeq}
\format{A data frame with 39997 rows and 7 variables:
\describe{
  \item{Marker.ID}{RefSeq name with its corrsponding gene symbol}
  \item{Chromosome}{1-22, X and Y}
  \item{Start}{starting position, in basepair number}
  \item{Stop}{ending position, in basepair number}
  \item{Strand}{positive or negative strand, in + or - symbols}
  \item{Gene.Symbol}{Gene name}
  \item{Transcript}{RefSeq name}
}}
\source{
UCSC Genome Browser:
  \url{http://hgdownload.cse.ucsc.edu/downloads.html}
}
\usage{
hg19DBNM
}
\value{
data.table
}
\description{
The human genome reference used here is RefSeq transcripts in version hg19
from UCSC Genome Browser. The transcripts with NM marker ID, which are
protein-codeing, were selected to be our reference database and provided as
hg19DBNM.rda.
}
\details{
This reference provides region information, including chromosome number,
starting position, ending position, strand and gene symbols, for converting
copy number alteration data into human genes.
}
\keyword{datasets}

